Move IDalamudAssetManager to correct namespace
