+++ Dalamud/Storage/Assets/DalamudAssetManager.cs
@@ -14,6 +14,7 @@ using Dalamud.Interface.Textures.TextureWraps.Internal;
 using Dalamud.IoC;
 using Dalamud.IoC.Internal;
 using Dalamud.Networking.Http;
+using Dalamud.Plugin.Services;
 using Dalamud.Utility;
 using Dalamud.Utility.Timing;
